add nouvelles fonctionnalités

+++ src/Controller/TaskController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Task;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
+use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
@@ -23,9 +24,19 @@ class TaskController extends AbstractController
     }
 
     /**
+     * @Route("/tasks/done", name="task_list_isDone")
+     */
+    public function listActionisDone(TaskRepository $taskRepository)
+    {
+        $tasks = $taskRepository->findBy(['isDone' => true]);
+
+        return $this->render('task/list.html.twig', ['tasks' => $tasks]);
+    }
+
+    /**
      * @Route("/tasks/create", name="task_create")
      */
-    public function createAction(Request $request)
+    public function createAction(Request $request, EntityManagerInterface $em)
     {
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
@@ -36,8 +47,6 @@ class TaskController extends AbstractController
             $task->setCreatedAt(new \DateTime(date('Y-m-d H:i:s')))
                 ->setIsDone(false)
                 ->setUser($this->getUser());
-
-            $em = $this->getDoctrine()->getManager();
 
             $em->persist($task);
             $em->flush();
@@ -53,17 +62,20 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/edit", name="task_edit")
      */
-    public function editAction(Task $task, Request $request)
+    public function editAction(Task $task, Request $request, EntityManagerInterface $em)
     {
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $em->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
+            if ($task->isDone() === true) {
+                return $this->redirectToRoute('task_list_isDone');
+            }
             return $this->redirectToRoute('task_list');
         }
 
@@ -76,11 +88,11 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/toggle", name="task_toggle")
      */
-    public function toggleTaskAction(Task $task)
+    public function toggleTaskAction(Task $task, EntityManagerInterface $em)
     {
         $task->toggle(!$task->isDone());
 
-        $this->getDoctrine()->getManager()->flush();
+        $em->flush();
 
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
 
@@ -90,14 +102,27 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/delete", name="task_delete")
      */
-    public function deleteTaskAction(Task $task)
+    public function deleteTaskAction(Task $task, EntityManagerInterface $em)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        $user = $this->getUser()->getRoles();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if ($task->getUser() === null && $user[0] == "ROLE_ADMIN") {
+            $em->remove($task);
+            $em->flush();
 
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        }
+
+        if ($task->getUser() == $this->getUser()) {
+            $em->remove($task);
+            $em->flush();
+
+            $this->addFlash('success', sprintf('La tâche %s a bien été supprimée.', $task->getTitle()));
+        }
+
+        if ($task->isDone() === true) {
+            return $this->redirectToRoute('task_list_isDone');
+        }
         return $this->redirectToRoute('task_list');
     }
 }
